Add option to select which neighbor to pick
If there are only two containers in the parent container, it is clear
which one you need to swap with. If there are more we need to decide
which one to pick and decent to to find the swapee.

This can be selected by using the `--neighbor` option which can be
either "first" (default) or "last". Selecting something in between seems
unfeasible.